requirements.txt
absl-py==1.0.0
addict==2.4.0
asttokens 
attrs==21.4.0
backcall 
batchgenerators==0.24
brotlipy==0.7.0
cachetools==5.1.0
certifi==2022.5.18.1
cffi 
charset-normalizer 
click==8.1.3
colorlover==0.3.0
contextlib2==21.6.0
cryptography 
cycler==0.11.0
debugpy 
decorator 
Deprecated==1.2.13
dicom2nifti==2.3.3
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints 
et-xmlfile==1.1.0
executing 
filelock==3.7.1
fonttools==4.33.3
fsspec==2022.5.0
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grad-cam==1.3.9
grpcio==1.46.3
h5py==3.7.0
hparams==0.3.0
huggingface-hub==0.8.1
humanize==4.1.0
idna 
imageio==2.19.3
importlib-metadata==4.11.4
iniconfig==1.1.1
iopath==0.1.9
ipykernel 
ipython 
jedi 
joblib==1.1.0
jupyter-client 
jupyter-core 
kiwisolver==1.4.2
linecache2==1.0.0
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline 
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
ml-collections==0.1.1
# mmcv==1.5.2
# mmcv-full==1.5.2
# mmdet==2.25.0
multilabel-metrics==0.0.1
nest-asyncio 
networkx==2.8.2
nibabel==3.2.2
nnunet==1.7.0
nose==1.3.7
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
parso 
pexpect 
pickleshare 
Pillow==9.0.1
plot-metric==0.0.6
pluggy==1.0.0
portalocker==2.4.0
prettytable==3.3.0
prompt-toolkit 
protobuf==3.19.0
ptyprocess 
pure-eval 
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser 
pydicom==2.3.0
Pygments 
pyOpenSSL 
pyparsing==3.0.9
PySocks 
pytest==7.1.2
python-dateutil 
pytorch-lightning==0.10.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
regex==2022.6.2
requests 
requests-oauthlib==1.3.1
resnest==0.0.6b20220613
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
self-attention-cv==1.2.3
setproctitle==1.2.3
SimpleITK==2.1.1.2
six 
sklearn==0.0
stack-data 
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
termcolor==1.1.0
terminaltables==3.1.10
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.6.2.dev0
tokenizers==0.12.1
tomli==2.0.1
torch==1.11.0
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchinfo==1.7.0
torchio==0.18.77
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.12.0
tornado 
tqdm==4.64.0
traceback2==1.4.0
traitlets 
transformers==4.20.1
ttach==0.0.3
typeguard==2.13.3
typing_extensions 
unittest2==1.1.0
urllib3 
wcwidth 
Werkzeug==2.1.2
wrapt==1.14.1
yacs==0.1.8
yapf==0.32.0
zipp==3.8.0
zmq==0.0.0